refactor(kernel): remove LLM-hang watchdog + .llm_hang sentinel
The 120s _send_with_watchdog timer false-fired on routine slow turns
(thinking models at high context regularly exceed 120s) and the
filesystem sentinel it wrote could outlive a successful refresh,
leaving relaunched agents stuck refusing to wake.

Removed:
- _on_llm_hang / _send_with_watchdog / _write/_remove/_exists/_age
  signal helpers (turn.py)
- _mark/_handle_worker_still_running and the sentinel-coupled AED
  branch (turn.py)
- Wake-refusal block + TTL recovery in _run_loop (turn.py)
- .llm_hang clear in _perform_refresh (lifecycle.py)
- .llm_hang entry in CPR signal-file sweep (lingtai/agent.py)
- Watchdog-era docstring on WorkerStillRunningError (llm_utils.py)
- Original design doc and sentinel-coupled tests

Preserved (the legitimate ChatInterface-safety property):
- WorkerStillRunningError still raises out of _send when the worker
  outlives timeout+grace
- The run loop now handles it inline: STUCK -> ASLEEP with
  skip_post_turn_save=True, no filesystem coupling

Net: turn.py 1145 -> 941 lines. 914 tests pass on the worktree
(1 unrelated flake in test_molt_notification_persistence under
parallel tmp-dir contention, reproduces on origin/main).
